+++ PublicAccessTV/src/Scene.cs
@@ -1,10 +1,10 @@
-﻿using Microsoft.Xna.Framework.Audio;
-using StardewModdingAPI;
+﻿using StardewModdingAPI;
 using StardewValley;
 using StardewValley.Objects;
 using System;
 using System.IO;
 using System.Media;
+using System.Runtime.CompilerServices;
 
 namespace PublicAccessTV
 {
@@ -18,15 +18,16 @@ namespace PublicAccessTV
 		public TemporaryAnimatedSprite background;
 		public TemporaryAnimatedSprite overlay;
 
-		public string soundCueName;
-		public string musicTrack = "none";
+		// Don't use looping sound cues or long external sound assets, since
+		// neither can be stopped at the end of scenes due to cross-platform
+		// limitations (Windows and Android, respectively). Music tracks are
+		// stopped at the appropriate time on all platforms.
+		public string musicTrack;
 		public string soundAsset;
+		public string soundCue;
 
 		public Action beforeAction;
 		public Action afterAction;
-
-		private ICue soundCue;
-		private SoundPlayer soundPlayer;
 
 		private TV currentTV;
 		private Channel currentChannel;
@@ -47,26 +48,22 @@ namespace PublicAccessTV
 			if (beforeAction != null)
 				beforeAction.Invoke ();
 
-			// Don't use looping sound cues for this, as they can't be stopped
-			// safely (see End method below).
-			if (soundCueName != null)
-			{
-				soundCue = Game1.soundBank.GetCue (soundCueName);
-				if (soundCueName == "distantTrain")
-					soundCue.SetVariable ("Volume", 100f);
-				soundCue.Play ();
-			}
-
 			Game1.changeMusicTrack (musicTrack ?? "none", false,
 				Game1.MusicContext.Event);
 
-			if (soundAsset != null)
+			if (soundAsset != null &&
+				Constants.TargetPlatform != GamePlatform.Android)
 			{
-				string soundPath = Path.Combine (Helper.DirectoryPath,
-					"assets", $"{soundAsset}.wav");
-				soundPlayer = new SoundPlayer (soundPath);
-				soundPlayer.Play ();
+				playCustomSound ($"{soundAsset}.wav");
 			}
+			else if (soundCue != null)
+			{
+				ICue cue = Game1.soundBank.GetCue (soundCue);
+				if (soundCue == "distantTrain")
+					cue.SetVariable ("Volume", 100f);
+				cue.Play ();
+			}
+
 
 			Helper.Reflection.GetField<TemporaryAnimatedSprite> (tv, "screen")
 				.SetValue (background);
@@ -79,17 +76,30 @@ namespace PublicAccessTV
 
 		private void end ()
 		{
-			// For some reason, certain Windows users do not have the method
-			// StardewValley.ICue.Stop defined. As such, do not attempt to
-			// stop any running vanilla sound cue here.
-				
-			if (soundPlayer != null)
-				soundPlayer.Stop ();
-
 			if (afterAction != null)
 				afterAction.Invoke ();
 			
 			currentChannel.runProgram (currentTV);
+		}
+
+		protected void playCustomSound (string filename)
+		{
+			string path = Path.Combine (Helper.DirectoryPath, "assets", filename);
+			try
+			{
+				playSoundWithSoundPlayer (path);
+			}
+			catch (TypeLoadException)
+			{
+				// best effort, since Android lacks System.Media
+			}
+		}
+
+		[MethodImplAttribute(MethodImplOptions.NoInlining)] 
+		private void playSoundWithSoundPlayer (string path)
+		{
+			SoundPlayer sound = new SoundPlayer (path);
+			sound.Play ();
 		}
 	}
 }